#169 fast fixes (#178)
* Catch the promises

* Validate booleans

* Update awaitMessage.js

* Update ping.js

* Update package.json

* Update CHANGELOG.md

+++ functions/awaitMessage.js
@@ -22,6 +22,6 @@ module.exports = async (client, msg, cmd, args, error) => {
       }
     } else return "Aborted";
   } finally {
-    message.delete();
+    if (message.deletable) message.delete();
   }
 };
